Rank iterator short overflow bugfix (#291)
* Regression tests for Peekable{Short,Int}RankIterators bug

* PeekableShortRankIterator.peekNextRank return type changed to int

* TestRankIteratorsOfContainers fix

+++ roaringbitmap/src/main/java/org/roaringbitmap/ArrayContainer.java
@@ -1326,8 +1326,8 @@ final class ArrayContainerShortIterator implements PeekableShortRankIterator {
   }
 
   @Override
-  public short peekNextRank() {
-    return (short) (pos + 1);
+  public int peekNextRank() {
+    return pos + 1;
   }
 
   @Override
